Se realizó una correción con respecto al acceso de la base de datos con la ruta proporcionada por el programa, para la misma se implementó la libreria "OS" y se creó un archivo 'conexion_base.py' para modularizar correctamente la conexión a la base. Se implementó el uso de un cursor para las consultas a modo de evitar cometer errores con el acceso a los datos de manera "local".
