--- api/db/hosteria_byteados.db.sqbpro
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><sqlb_project><db path="/media/lucas/Disco secundario/Archivos-PERSONALES-COMPARTIDOS/PROYECTOS PROGRAMACION/Intro-Des-Soft/byteados/UBA_TP_IDS/api/db/hosteria_byteados.db" readonly="0" foreign_keys="1" case_sensitive_like="0" temp_store="0" wal_autocheckpoint="1000" synchronous="2"/><attached/><window><main_tabs open="structure browser pragmas query" current="0"/></window><tab_structure><column_width id="0" width="300"/><column_width id="1" width="0"/><column_width id="2" width="100"/><column_width id="3" width="3227"/><column_width id="4" width="0"/><expanded_item id="0" parent="1"/><expanded_item id="0" parent="0"/><expanded_item id="1" parent="0"/><expanded_item id="2" parent="0"/><expanded_item id="1" parent="1"/><expanded_item id="2" parent="1"/><expanded_item id="3" parent="1"/></tab_structure><tab_browse><current_table name="4,12:mainhabitaciones"/><default_encoding codec=""/><browse_table_settings/></tab_browse><tab_sql><sql name="SQL 1">/* utilizar desde acá si se carga en sqlite */
-create table habitaciones(
-    habitacion_id varchar(15) not null,
-    habitacion_nombre varchar(30),
-    habitacion_descripcion varchar(500),
-    habitacion_precio_dia float,
-    constraint pk_habitacion_id primary key (habitacion_id)
-);
-
-create table reservas(
-    reserva_codigo int not null, /*si una PK es int, auto-increment se implementa automaticamente en sqlite*/
-    habitacion_id int not null,
-    reserva_fecha_ent date not null,
-    reserva_fecha_sal date not null,
-    reserva_id_cliente varchar(9) not null, /* DNI o Pasaporte */
-    reserva_telefono_cliente int not null,
-    reserva_mail_cliente int not null,
-    reserva_precio_total float not null,
-    constraint pk_reserva_codigo primary key (reserva_codigo),
-    constraint fk_habitacion_id FOREIGN KEY (habitacion_id) REFERENCES habitaciones(habitacion_id)
-);
-
-create table imagenes(
-    imagen_link varchar(2048) not null,
-    habitacion_id int,
-    imagen_descripcion varchar(100),
-    constraint pk_imagen_link primary key (imagen_link),
-    constraint fk_habitacion_id foreign key (habitacion_id) references habitaciones(habitacion_id)
-);</sql><current_tab id="0"/></tab_sql></sqlb_project>
